Kostilling
Java does not support passing int by reference. That's why class UserID created as a part of UserManager, and now it's main function take as parameter object of that type. Actualy this is axact defenition of "Side Effect", so this solution is kinda dirty, but otherwise we have to make UserID a memeber of UserManager, which is even worse, becouse we will no only have side effect (Even though it's on inner field) but also a time dependency.